Add files via upload
this was my version of the code. I did have to check my version with the one on udemy by the tutor. 

I did make my own version of the cents calculator which she made in with a lot less lines of code. but overall pretty happy with how it turned out